refactor: rename test files to descriptive names
week1_day1.rs -> memtable.rs
week1_day2.rs -> iterators.rs
week1_day3.rs -> block.rs
week1_day4.rs -> sst.rs
week1_day5.rs -> merge_iterator.rs
week1_day6.rs -> scan_flush.rs
week1_day7.rs -> bloom_compression.rs
week2_day1.rs -> compaction.rs
week2_day2.rs -> tiered_compaction.rs
week2_day3.rs -> leveled_compaction.rs
week2_day4.rs -> simple_leveled_compaction.rs
week2_day5.rs -> compaction_integration.rs
week2_day6.rs -> compaction_integration_2.rs
