+++ kv-engine/src/tests/compaction.rs
@@ -2,7 +2,7 @@ use std::{ops::Bound, path::Path, sync::Arc};
 
 use bytes::Bytes;
 use tempfile::tempdir;
-use week2_day1::harness::construct_merge_iterator_over_storage;
+use harness::construct_merge_iterator_over_storage;
 
 use self::harness::{check_iter_result_by_key, check_lsm_iter_result_by_key, sync};
 use super::*;
